removing viper
